Fix issuse 2
Add a widget to support the match level setting, update methods
to be able to work with that value, but do not add this field
if the value is not filled-in as it is normal with needles.
Also fix the provideProperties() that was failing when no
properties were defined in the json. Now, we can open the needle
even without those field set and we will recreate the proper json
when the needle is saved.